+++ depthai_ros_driver/cfg/parameters.cfg
@@ -82,7 +82,7 @@ camera.add("rgb_i_board_socket_id", int_t, 0, "Sensor board socket id", 0)
 camera.add("rgb_i_calibration_file", str_t, 0, "Path to calibration file", "")
 camera.add("rgb_i_fps", double_t, 0, "FPS", 60.0)
 camera.add("rgb_i_height", int_t, 0, "Image height", 720)
-camera.add("rgb_i_interleaved", bool_t, 0, "Is frame interleaved or planar", True)
+camera.add("rgb_i_interleaved", bool_t, 0, "Is frame interleaved or planar", False)
 camera.add("rgb_r_keep_preview_aspect_ratio", bool_t, 0, "Keep preview aspect ratio", True)
 camera.add("rgb_i_low_bandwidth", bool_t, 0, "Use encoding for data", False)
 camera.add("rgb_i_low_bandwidth_quality", int_t, 0, "Quality when using low-bandwidth mode", 50, 1, 100)
@@ -116,6 +116,9 @@ camera.add("stereo_i_decimation_filter_decimation_factor", int_t, 0, "Decimation
 camera.add("stereo_i_decimation_filter_decimation_mode", str_t, 0, "Decimation mode", "PIXEL_SKIPPING")
 camera.add("stereo_i_depth_filter_size", int_t, 0, "Depth filter size", 5)
 camera.add("stereo_i_depth_preset", str_t, 0, "Depth preset", "HIGH_ACCURACY")
+camera.add("stereo_i_enable_brightness_filter", bool_t, 0, "Enable brightness filter", False)
+camera.add("stereo_i_brightness_filter_min_brightness", int_t, 0, "Brightness filter min", 0)
+camera.add("stereo_i_brightness_filter_max_brightness", int_t, 0, "Brightness filter max", 256)
 camera.add("stereo_i_enable_companding", bool_t, 0, "Enable companding", False)
 camera.add("stereo_i_enable_decimation_filter", bool_t, 0, "Enable decimation filter", False)
 camera.add("stereo_i_enable_distortion_correction", bool_t, 0, "Enable distortion correction", True)
@@ -129,6 +132,12 @@ camera.add("stereo_i_low_bandwidth_quality", int_t, 0, "Quality when using low-b
 camera.add("stereo_i_lr_check", bool_t, 0, "Enable left-right check", True)
 camera.add("stereo_i_lrc_threshold", int_t, 0, "Left-right check threshold", 10)
 camera.add("stereo_i_max_q_size", int_t, 0, "Internal queue size", 30)
+camera.add("stereo_i_publish_left_rect", bool_t, 0, "Publish rectified left", False)
+camera.add("stereo_i_left_rect_low_bandwidth", bool_t, 0, "Low bandwidth for rectified left", False)
+camera.add("stereo_i_left_rect_low_bandwidth_quality", int_t, 0, "Low bandwidth left quality", 50)
+camera.add("stereo_i_publish_right_rect", bool_t, 0, "Publish rectified right", False)
+camera.add("stereo_i_right_rect_low_bandwidth", bool_t, 0, "Low bandwidth for rectified right", False)
+camera.add("stereo_i_right_rect_low_bandwidth_quality", int_t, 0, "Low bandwidth right quality", 50)
 camera.add("stereo_i_rectify_edge_fill_color", int_t, 0, "Rectify edge fill color", 0)
 camera.add("stereo_i_spatial_filter_alpha", double_t, 0, "Spatial filter alpha", 0.5)
 camera.add("stereo_i_spatial_filter_delta", int_t, 0, "Spatial filter delta", 20)
