+++ depthai_ros_driver/include/depthai_ros_driver/dai_nodes/nn/spatial_detection.hpp
@@ -47,9 +47,19 @@ class SpatialDetection : public BaseNode {
     void setupQueues(std::shared_ptr<dai::Device> device) override {
         nnQ = device->getOutputQueue(nnQName, ph->getParam<int>("i_max_q_size"), false);
         auto tfPrefix = getTFPrefix("rgb");
+        int width;
+        int height;
+        if(ph->getParam<bool>("i_disable_resize")) {
+            width = ph->getOtherNodeParam<int>("rgb", "i_preview_size");
+            height = ph->getOtherNodeParam<int>("rgb", "i_preview_size");
+        }
+        else{
+            width = imageManip->initialConfig.getResizeConfig().width;
+            height = imageManip->initialConfig.getResizeConfig().height;
+        }
         detConverter = std::make_unique<dai::ros::SpatialDetectionConverter>(tfPrefix + "_camera_optical_frame",
-                                                                             imageManip->initialConfig.getResizeConfig().width,
-                                                                             imageManip->initialConfig.getResizeConfig().height,
+                                                                             width,
+                                                                             height,
                                                                              false,
                                                                              ph->getParam<bool>("i_get_base_device_timestamp"));
         nnQ->addCallback(std::bind(&SpatialDetection::spatialCB, this, std::placeholders::_1, std::placeholders::_2));
@@ -62,8 +72,8 @@ class SpatialDetection : public BaseNode {
             ptInfoMan->setCameraInfo(sensor_helpers::getCalibInfo(*ptImageConverter,
                                                                   device,
                                                                   dai::CameraBoardSocket::RGB,
-                                                                  imageManip->initialConfig.getResizeWidth(),
-                                                                  imageManip->initialConfig.getResizeWidth()));
+                                                                  width,
+                                                                  height));
 
             ptPub = it.advertiseCamera(getName() + "/passthrough/image_raw", 1);
             ptQ->addCallback(std::bind(sensor_helpers::imgCB, std::placeholders::_1, std::placeholders::_2, *ptImageConverter, ptPub, ptInfoMan));
@@ -80,9 +90,8 @@ class SpatialDetection : public BaseNode {
             ptDepthQ = device->getOutputQueue(ptDepthQName, ph->getParam<int>("i_max_q_size"), false);
             ptDepthImageConverter = std::make_unique<dai::ros::ImageConverter>(tfPrefix + "_camera_optical_frame", false);
             ptDepthInfoMan = std::make_shared<camera_info_manager::CameraInfoManager>(ros::NodeHandle(getROSNode(), getName()), "/" + getName());
-            int width, height;
-            getROSNode().getParam("stereo_i_width", width);
-            getROSNode().getParam("stereo_i_height", height);
+            int width = ph->getOtherNodeParam<int>("stereo", "i_width");
+            int height = ph->getOtherNodeParam<int>("stereo", "i_height");
             ptDepthInfoMan->setCameraInfo(sensor_helpers::getCalibInfo(*ptDepthImageConverter, device, socket, width, height));
 
             ptDepthPub = it.advertiseCamera(getName() + "/passthrough_depth/image_raw", 1);
