+++ depthai_ros_driver/package.xml
@@ -32,9 +32,11 @@
   <depend>depthai_ros_msgs</depend>
   <depend>depthai_examples</depend>
   <depend>nodelet</depend>
+  <depend>pluginlib</depend>
 
   <export>
     <nodelet plugin="${prefix}/nodelet_plugins.xml" />
+    <depthai_ros_driver plugin="${prefix}/plugins.xml" />
     <build_type>catkin</build_type>
   </export>
 </package>
